refactor(layout): refactor naming and fix hydration miss matches
