+++ layouts/default.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { chakra, useColorModeValue } from '@chakra-ui/vue-next';
-import TopNavigation from '~/components/navigation/top-navigation.vue';
-import Sidebar from '~/components/navigation/sidebar.vue';
+import TheNavbar from '~/components/TheNavbar.vue'
+import TheSidebar from '~/components/navigation/TheSidebar.vue';
 
 /**
  * Styling
@@ -11,13 +11,15 @@ const color = useColorModeValue('gray.700', 'white');
 
 <template>
   <chakra.div :min-h="['auto', 'auto', '100vh']" w="100%" :color="color">
-    <TopNavigation />
+    <TheNavbar />
     <chakra.div max-w="8xl" mx="auto" d="flex">
       <!-- Sidebar Navigation -->
-      <chakra.div :display="{ base: 'none', lg: 'block' }" position="fixed" z-index="30" bottom="0" top="6rem"
+      <chakra.div 
+        :display="{ base: 'none', lg: 'block' }" position="fixed" z-index="30" bottom="0" top="6rem"
         left="max(0px, calc(50% - 45rem))" right="auto" width="19.5rem" pb="10" px="8" overflow-y="auto"
-        overscroll-behavior="contain">
-        <sidebar />
+        overscroll-behavior="contain"
+      >
+        <TheSidebar />
       </chakra.div>
       <chakra.main :pl="{ base: 4, lg: '19.5rem' }" pt="4" pb="24" :pr="{ base: 4, xl: 16 }" class="chakra-prose">
         <chakra.div :mr="{ xl: '15.5rem' }" :pl="{ base: 2, lg: '1rem' }">
